Move VersionUtils.ts into own package because imported at very low level
