+++ kit/ag-grid/index.ts
@@ -6,7 +6,7 @@
  */
 
 import {logError} from '@xh/hoist/utils/log';
-import {checkVersion} from '@xh/hoist/utils/js/VersionUtils';
+import {checkVersion} from '@xh/hoist/utils/version';
 
 /**
  * The exports below are ag-Grid components provided at runtime by applications.
