+++ svc/EnvironmentService.ts
@@ -11,7 +11,8 @@ import {action, makeObservable, observable} from '@xh/hoist/mobx';
 import hoistPkg from '@xh/hoist/package.json';
 import {Timer} from '@xh/hoist/utils/async';
 import {MINUTES, SECONDS} from '@xh/hoist/utils/datetime';
-import {checkMaxVersion, checkMinVersion, deepFreeze} from '@xh/hoist/utils/js';
+import {checkMaxVersion, checkMinVersion} from '@xh/hoist/utils/version';
+import {deepFreeze} from '@xh/hoist/utils/js';
 import {defaults, isFinite} from 'lodash';
 import mobxPkg from 'mobx/package.json';
 import {version as reactVersion} from 'react';
