+++ utils/version/index.ts
@@ -4,8 +4,5 @@
  *
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
-export * from './HtmlUtils';
-export * from './LangUtils';
-export * from './Decorators';
-export * from './DomUtils';
-export * from './TestUtils';
+
+export * from './VersionUtils';
